chore: reorg eslint rules into separate files
