+++ projenrc/eslint/jsdoc.ts
@@ -1,0 +1,7 @@
+// JSDoc
+export default {
+  'jsdoc/require-param-description': ['error'],
+  'jsdoc/require-property-description': ['error'],
+  'jsdoc/require-returns-description': ['error'],
+  'jsdoc/check-alignment': ['error'],
+};
